+++ var10/lab8/Wood.h
@@ -3,7 +3,7 @@
 
 
 struct Wood {
-    double data;
+    int data;
     Wood* left;
     Wood* right;
 };
